Move to normalization folder
