+++ include/ck/tensor_operation/gpu/grid/normalization/gridwise_normalization_naive_variance.hpp
